Renamed unrelated.py to multi_theme.py.
Also sticking with "primary" as the label of the first theme. Replacing
"root" with that for consistency.
