+++ tests/unit_tests/test_docs/test-multi-theme/test_parse_themes.py
@@ -1,7 +1,7 @@
 """Tests."""
 import pytest
 
-from sphinx_carousel.unrelated import MultiTheme
+from sphinx_carousel.multi_theme import MultiTheme
 
 
 def test():
